// frontend/src/dtos/user/challenge/user-challenge-details.dto.ts
import {
  ChallengeAccessType,
  ChallengeCategory,
  ChallengeDifficulty,
} from "@/types/challenge/challenge.types";

import {
  ChallengeActivityType,
  ChallengeActivityValueType,
} from "@/types/challenge/challenge-day.types";

import { UserChallengeStatus } from "@/types/user/challenge/user-challenge.types";

export interface UserChallengeActivityDTO {
  id: string;
  type: ChallengeActivityType;
  title: string;
  description?: string;
  instructions?: string;
  valueType: ChallengeActivityValueType;
  targetValue?: number;
  unit?: string;
  estimatedDurationMinutes?: number;
  imageUrl?: string;
  videoUrl?: string;
  isRequired: boolean;
  order: number;
  completed: boolean;
}

export interface UserChallengeDayDTO {
  id: string;
  dayNumber: number;
  title?: string;
  description?: string;
  activities: UserChallengeActivityDTO[];
}

export interface UserChallengeProgressDTO {
  id: string;
  challengeDayId: string;
  activityId: string;
  completedAt: Date;
}

export interface UserChallengeDetailsDTO {
  id: string;
  status: UserChallengeStatus;
  joinedAt: Date;
  startedAt: Date;
  currentDay: number;
  progressPercentage: number;
  currentStreak: number;
  longestStreak: number;
  completedAt?: Date;

  challenge: {
    id: string;
    title: string;
    description: string;
    instructions?: string;
    coverImageUrl?: string;
    category: ChallengeCategory;
    difficulty: ChallengeDifficulty;
    accessType: ChallengeAccessType;
    durationDays: number;
  };

  days: UserChallengeDayDTO[];
  progress: UserChallengeProgressDTO[];
}